+++ packages/@baldeweg/ui/src/components/BNav/BNav.stories.js
@@ -9,7 +9,7 @@ export default {
       description: 'Navigation items array',
     },
   },
-  tags: ['experimental'],
+  tags: ['deprecated'],
 }
 
 const Template = (args) => ({